refactor: language filter per review – remove getJobLanguage, closed word sets, schema in create_tables, ERD update
