+++ client/src/pages/openPositions/OpenPositions.jsx
@@ -23,7 +23,6 @@ import createSortComparator from "../../util/createSortComparator";
 import { DELAYED_CLEAR_INTERVAL } from "../../util/constants";
 import { findFilterOptions, filterJobs } from "../../util/filterJobs";
 import getSkillsInDescription from "../../util/getSkillsInDescription";
-import detectLanguage from "shared/utils/detectLanguage.js";
 // Assets
 import { gif } from "../../assets/index.js";
 // Styles
@@ -93,13 +92,9 @@ export default function OpenPositions() {
         job.normalized_description || "",
         skills,
       );
-      const language =
-        job.language != null && job.language !== ""
-          ? job.language
-          : detectLanguage(job.normalized_description ?? "");
       return {
         ...job,
-        language,
+        language: job.language,
         skillsInDescription,
         skillsMatch: String(skillsInDescription.length).padStart(2, "0"),
       };
